+++ zoo/squid/mysql/statement.cpp
@@ -208,7 +208,7 @@ std::uint64_t statement::affected_rows()
 
 /*static*/ void statement::execute(MYSQL& connection, const std::string& query)
 {
-	if (0 != mysql_real_query(&connection, query.c_str(), query.length()))
+	if (0 != mysql_real_query(&connection, query.c_str(), static_cast<unsigned long>(query.length())))
 	{
 		ZOO_THROW_EXCEPTION(error{ "mysql_real_query failed", connection });
 	}
